requirements.txt
langchain_google_genai
langchain
spacy
streamlit
nltk
gensim
wordcloud
python-dotenv